+++ forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, TextAreaField
+from wtforms import StringField, TextAreaField, SelectMultipleField
 from wtforms.fields.simple import SubmitField, PasswordField
 from wtforms.validators import DataRequired, Email, EqualTo, Length
 
@@ -27,4 +27,5 @@ class PostForm(FlaskForm):
     """Форма для создания поста"""
     title = StringField('Title', validators=[DataRequired(), Length(max=150)])
     content = TextAreaField('Content', validators=[DataRequired()])
+    tags = SelectMultipleField('Tags', choices=[('nature', 'Природа'), ('birds', 'Птицы'), ('animals', 'Животные')])
     submit = SubmitField('Создать пост')
