+++ models/__init__.py
@@ -1,3 +1,3 @@
 from .database import db
 from .user import User
-from .users_post import Post
+from .post_and_tags import Post, Tag, PostTag
