+++ templates/base.html
@@ -3,12 +3,20 @@
 <head>
     <meta charset="UTF-8">
     <title>
-        {% block title %}
-        {% endblock %}
+        {% block title %}{% endblock %}
     </title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
           integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link href="https://fonts.googleapis.com/css2?family=Raleway:wght@400;700&display=swap" rel="stylesheet">
+    <style>
+        .badge {
+            z-index: 1;
+        }
+
+        .card-title {
+            margin-top: 1rem;
+        }
+    </style>
 </head>
 <body>
 <header class="p-3 text-bg-dark">
@@ -21,7 +29,7 @@
             </a>
 
             <ul class="nav col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0">
-                <li><a href="/" class="nav-link px-2 text-white">Главная</a></li>
+                <li><a href="/" class="nav-link px-2 text-secondary">Главная</a></li>
                 <li><a href="/blog" class="nav-link px-2 text-white">Блог</a></li>
             </ul>
 
@@ -29,11 +37,7 @@
             <div class="text-end">
                 {% if current_user.is_authenticated %}
                 <span class="text-white me-2">Аккаунт, {{ current_user.username }}</span>
-
-                <!--Эта строка проверяет, находится ли текущий путь (URL) не на странице личного кабинета-->
-                {% if request.path != url_for('personal_account') %}
                 <a href="{{ url_for('personal_account') }}" class="btn btn-outline-warning">Личный кабинет</a>
-                {% endif %}
                 <a href="{{ url_for('logout') }}" class="btn btn-warning">Выйти</a>
                 {% else %}
                 <a href="{{ url_for('login') }}" class="btn btn-outline-light me-2">Войти в личный кабинет</a>
