+++ templates/blog.html
@@ -8,8 +8,8 @@
 <br>
 <h3>Блог об экзотических птицах, их видовом разнообразии и жизни в естественной среде</h3>
 <br>
-<h3>Изучите удивительные особенности, поведение и уникальные адаптации, которые позволяют
-    этим ярким существам процветать в самых разных уголках нашей планеты.</h3>
+<h2>Изучите удивительные особенности, поведение и уникальные адаптации, которые позволяют
+    этим ярким существам процветать в самых разных уголках нашей планеты.</h2>
 <br>
 {% if posts %}
 <div class="row">
@@ -17,11 +17,16 @@
     {% if post.is_published %}
     <div class="col-md-4">
         <div class="card mb-4 shadow-sm">
-            <div class="card-body">
+            <div class="card-body position-relative">
                 <div class="d-flex justify-content-between align-items-center">
-                    <h5 class="card-title">{{ post.title }}</h5>
-                    <span class="badge bg-warning position-absolute top-0 end-0 m-2">{{ post.user.username }}</span>
+                    <span class="badge bg-warning">{{ post.user.username }}</span>
+                    <div>
+                        {% for tag in post.tags %}
+                        <span class="badge bg-info">{{ tag.name }}</span>
+                        {% endfor %}
+                    </div>
                 </div>
+                <h5 class="card-title mt-3">{{ post.title }}</h5>
                 <p class="card-text">{{ post.content[:100] }}...</p>
                 <a href="{{ url_for('view_post', post_id=post.id) }}" class="btn btn-success">Читать далее</a>
             </div>
@@ -34,4 +39,5 @@
 <div class="alert alert-info">Пока в этом блоге ничего не опубликовано.</div>
 {% endif %}
 <a href="/" class="btn btn-secondary">Вернуться на главную</a>
+
 {% endblock %}
